Q1/requirements.txt
einops>=0.6.1
matplotlib>=3.7.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
PyYAML>=6.0
torch>=2.2.0,<2.7.0
torchvision>=0.17.0,<0.22.0
tqdm>=4.66.0
shapely>=2.0.0
Pillow>=9.5.0